requirements.txt
django>=3.1,<3.2
djangorestframework>=3.12,<3.13
psycopg2-binary==2.9.1
